+++ packages/ui/src/lib/table/table-cell-def.ts
@@ -1,0 +1,12 @@
+import { Directive, inject, TemplateRef } from '@angular/core';
+
+export interface ElTableCellContext {
+  $implicit: object;
+}
+
+@Directive({
+  selector: 'ng-template[elTableCell]',
+})
+export class ElTableCell {
+  readonly template = inject(TemplateRef<ElTableCellContext>);
+}